src/main/java/cz/gattserver/grass/hw/interfaces/HWItemOverviewTO.java
package cz.gattserver.grass.hw.interfaces;

import java.io.Serializable;
import java.math.BigDecimal;
import java.time.LocalDate;

/**
 * HW Objekt
 */
public class HWItemOverviewTO implements Serializable {

	private static final long serialVersionUID = 3678406951423588173L;

	/**
	 * Identifikátor hw
	 */
	private Long id;

	/**
	 * Název
	 */
	private String name;

	/**
	 * Datum zakoupení (získání)
	 */
	private LocalDate purchaseDate;

	/**
	 * Cena
	 */
	private BigDecimal price;

	/**
	 * Stav hw - funkční, rozbitý, poruchový, bližší popis
	 */
	private HWItemState state;

	/**
	 * Součást celku
	 */
	private String usedInName;

	/**
	 * Spravováno pro (spravuju tohle zařízení někomu?)
	 */
	private String supervizedFor;

	/**
	 * Je položka veřejně viditelná?
	 */
	private Boolean publicItem;

	public String getSupervizedFor() {
		return supervizedFor;
	}

	public void setSupervizedFor(String supervizedFor) {
		this.supervizedFor = supervizedFor;
	}

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public LocalDate getPurchaseDate() {
		return purchaseDate;
	}

	public void setPurchaseDate(LocalDate purchaseDate) {
		this.purchaseDate = purchaseDate;
	}

	public BigDecimal getPrice() {
		return price;
	}

	public void setPrice(BigDecimal price) {
		this.price = price;
	}

	public HWItemState getState() {
		return state;
	}

	public void setState(HWItemState state) {
		this.state = state;
	}

	public String getUsedInName() {
		return usedInName;
	}

	public void setUsedInName(String usedInName) {
		this.usedInName = usedInName;
	}

	public Boolean getPublicItem() {
		return publicItem;
	}

	public void setPublicItem(Boolean publicItem) {
		this.publicItem = publicItem;
	}

	@Override
	public int hashCode() {
		return id == null ? super.hashCode() : id.hashCode();
	}

	@Override
	public boolean equals(Object obj) {
		if (obj == null)
			return false;
		if (obj instanceof HWItemOverviewTO) {
			HWItemOverviewTO hw = (HWItemOverviewTO) obj;
			if (getId() != null)
				return getId().equals(hw.getId());
			return super.equals(hw);
		}
		return false;
	}

}
